Add requires for backward compat
