--- ext/SimpleBatchedNonlinearSolveExt/SimpleBatchedNonlinearSolveExt.jl
@@ -1,9 +1,0 @@
-module SimpleBatchedNonlinearSolveExt
-
-using SimpleNonlinearSolve, SciMLBase, NNlib
-
-include("utils.jl")
-include("broyden.jl")
-include("lbroyden.jl")
-
-end
